requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
openai==1.106.1
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
gunicorn==21.2.0
